[VHD Backups] Add WSUS WID optional restart
If WID is being used as the database backend for WSUS on Windows Server
2012 R2, then there seems to be an increase of memory as time goes on.
This can cause VMs that were previous on to not start due to insuffient
memory. The solution is to restart the WID service prior to restarting
the VMs. Note that this will briefly impact WSUS while the WID backend
restart occurs.